+++ src/main/java/com/finalproj/orbitflow/attendance/attendanceDefaultRule/repository/AttendanceRuleRepository.java
@@ -1,6 +1,6 @@
-package com.finalproj.orbitflow.attendance.attendanceRule.repository;
+package com.finalproj.orbitflow.attendance.attendanceDefaultRule.repository;
 
-import com.finalproj.orbitflow.attendance.attendanceRule.entity.AttendanceRule;
+import com.finalproj.orbitflow.attendance.attendanceDefaultRule.entity.AttendanceRule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
