+++ src/main/java/com/finalproj/orbitflow/attendance/attendanceExceptionRule/dto/EmpAttRuleUpdateReqDto.java
@@ -1,4 +1,4 @@
-package com.finalproj.orbitflow.attendance.employeeAttRule.dto;
+package com.finalproj.orbitflow.attendance.attendanceExceptionRule.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -7,16 +7,18 @@ import java.time.LocalTime;
  * Please explain the class!!!
  *
  * @author : rlagkdus
- * @filename : EmpAttRuleCreateReqDto
+ * @filename : EmpAttRuleUpdateReqDto
  * @since : 2025. 12. 18. 목요일
  */
-// 규칙 생성 요청 DTO (POST /exception)
-public record EmpAttRuleCreateReqDto(
-        Long employeeId,
+
+// 규칙 수정 요청 DTO (PUT /exception/{ruleId})
+public record EmpAttRuleUpdateReqDto(
         LocalTime startTime,
         LocalTime endTime,
         Integer breakMinutes,
         String reason,
         LocalDate validFrom,
-        LocalDate validTo
-) { }
+        LocalDate validTo,
+        Boolean isActive
+) {
+}
