+++ src/main/java/com/finalproj/orbitflow/attendance/commute/service/CommuteService.java
@@ -3,12 +3,12 @@ package com.finalproj.orbitflow.attendance.commute.service;
 import com.finalproj.orbitflow.attendance.commute.dto.ActiveRuleResDto;
 import com.finalproj.orbitflow.attendance.commute.dto.TodayAttResDto;
 import com.finalproj.orbitflow.attendance.commute.entity.Attendance;
-import com.finalproj.orbitflow.attendance.attendanceRule.entity.AttendanceRule;
-import com.finalproj.orbitflow.attendance.employeeAttRule.entity.EmployeeAttRule;
+import com.finalproj.orbitflow.attendance.attendanceDefaultRule.entity.AttendanceRule;
+import com.finalproj.orbitflow.attendance.attendanceExceptionRule.entity.EmployeeAttRule;
 import com.finalproj.orbitflow.attendance.commute.enums.AttendanceStatus;
 import com.finalproj.orbitflow.attendance.commute.repository.AttendanceRepository;
-import com.finalproj.orbitflow.attendance.attendanceRule.repository.AttendanceRuleRepository;
-import com.finalproj.orbitflow.attendance.employeeAttRule.repository.EmployeeAttRuleRepository;
+import com.finalproj.orbitflow.attendance.attendanceDefaultRule.repository.AttendanceRuleRepository;
+import com.finalproj.orbitflow.attendance.attendanceExceptionRule.repository.EmployeeAttRuleRepository;
 import com.finalproj.orbitflow.hr.employee.entity.Employee;
 import com.finalproj.orbitflow.hr.employee.enums.WorkStatus;
 import com.finalproj.orbitflow.hr.employee.repository.EmployeeRepository;
