+++ src/main/resources/templates/fragments/header.html
@@ -11,7 +11,7 @@
         <a th:href="@{/view/approval}"
            th:classappend="${currentGNB}=='approval' ? ' selected' : ''">결재</a>
 
-        <a th:href="@{/view/work}"
+        <a th:href="@{/view/attendance/commute}"
            th:classappend="${currentGNB}=='work' ? ' selected' : ''">근태</a>
 
         <a th:href="@{/view/schedule}"
